server

Module action

Module for game actions like playing a card or scoring hands.

Functions

cut_for_deal

Performs the “cut for deal” action for a user in a specific game.

discard_cards_to_crib

Discards a set of cards from a player’s hand to the crib in a game.

go

Performs a “go” action for the user in a pegging phase of the game.

host_game

Hosts a new game for the specified user.

join_game

Joins an existing game as a guest.

play_card

Plays a card for the user in the pegging phase of a game.

play_computer

Starts a game against the computer for the specified user.

score_crib

Acknowledges the current dealer score and goes on to score the crib for the specified user if both players have acknowledged.

score_dealer

Acknowledges the current pone score and goes on to score the dealer’s hand for the specified user if both players have acknowledged.

score_pone

Acknowledges the end of plays and goes on to score the pone’s hand for the specified user if both players have acknowledged.

start_game

Acknowledges the cuts for deal, then starts a hosted game for the specified user.

start_next_round

Acknowledges the crib score and goes on start the next deal / discarding round if both players have acknowledged.