Module server::api::v1::handlers::demos

Demo endpoints

Structs

demos

GET endpoint to return demo information.

demos_add

POST endpoint to upload a new demo changelog entry. Returns the new demo ID.

demos_changelog

Accepts field values for both a changelog, and a demo file.

demos_delete

DELETE endpoint to remove a demo from both backbalze and the database.

Functions

add_to_database 🔒

Adds a demo and changelog insert to the database.

b2_client_and_auth 🔒

Returns a client, and an authenticated session for use with backblaze.

delete_demo_db 🔒

Once the file has been removed, delete the demo entry.

delete_demo_file 🔒

Deletes the demo from backblaze.

generate_file_name 🔒

Create file_name

get_changelog_and_demo_id 🔒

Takes in either a demo_id or a changelog_id, and returns a changelog entry and a demno_id.

parse_and_write_multipart 🔒

Helper function that handles parsing the multipart and writing the file out locally

upload_demo 🔒

Handles uploading the demo file.