Module server::api::v1::handlers::users

User-related endpoints.

Structs

avatar_update

PUT method to update the avatar for a user in the database.

banned_user

GET method to return a bool based on if a user is banned or not.

banned_users_all

GET method to get all profile_numbers of all banned users on the board.

donators

GET method to return all user information for donators on the boards.

profile

GET method for most of a given user’s profile information.

user

GET method for user information for a specific profile_number.

user_add

POST method to upload a new user to the boards.

wall_of_shame

GET method to return UsersDisplay for all banned users on the boards.

Functions

profile_from_cache

Pulls & clones the data from the ranks cache to be used for the profile endpoint.