Function server::auth::find_user

pub fn find_user(_name: &str, _passwd: &str) -> Result<User, AuthError>

Validates username and password and returns the matched user.

Note: Currently nothing is checked yet and a meaningless User object is returned!

Failures

If the user was not found or the password does not match, an Err value is returned. See AuthError for more information.