server::conexiones::handler_cliente

Struct HandlerCliente

pub struct HandlerCliente<R, W>
where
    R: BufRead,
    W: Write,
{
    reader_stream: R,
    writer_stream: W,
    counter_calculadora: Arc<Mutex<Calculadora>>,
    termino_conexion: bool,
    buffer: String,
}

Maneja la conexion de un cliente en particular.

Fields
reader_stream: R

Canal por el cual se leeran las peticiones del cliente.

writer_stream: W

Canal por el cual se le enviaran repuestas al cliente.

counter_calculadora: Arc<Mutex<Calculadora>>

Calculadora sobre la que se aplicaran las peticiones del cliente.

termino_conexion: bool

Indica si el cliente termino o no de enviar peticiones

Un cliente termina de enviar peticiones cuando envia el mensaje Get.

buffer: String

En donde se almacenaran los mensajes enviados por el cliente.

Implementations

impl<R, W> HandlerCliente<R, W>
where R: BufRead, W: Write,

pub fn new( reader_stream: R, writer_stream: W, counter_calculadora: Arc<Mutex<Calculadora>>, buffer: String, ) -> Self

Crea y devuelve una nueva instancia de HandlerCliente.

pub fn gestionar(&mut self) -> Result<(), ErrorEnThread>

Se encarga del manejo general del cliente.

Permanece en un loop hasta que el cliente cierre la conexion de forma abrupta o hasta que ya no tenga que leer.

Lee la peticion del cliente, la procesa y genera una respuesta adecuada para enviarsela.

fn leer_mensaje(&mut self) -> Result<usize, ErrorEnThread>

Lee el mensaje que el cliente le envio por medio del reader_stream. En caso de que la conexion aun no deberia haber terminado pero igualmente no haya que leer, o si hay algun error durante la lectura se retorna un ErrorEnThread

fn procesar_mensaje(&mut self) -> Result<Option<MensajeCliente>, ErrorEnThread>

Procesa el mensaje enviado por el cliente, en caso de que este sea valido se retorna un MensajeCliente segun la finalidad de este (dentro de un Option). Si el mensaje es invalido se lo notifica al cliente y devuelve un None.

Si hay algun error durante este procesamiento, por ejemplo, que el mensaje enviado por el cliente no sea un mensaje que el servidor sepa responder se le envia un mensaje al cliente especificandole esto. Y si durante el envio de este mensaje hay algun inconveniente se retorna un ErrorEnThread

fn ejecutar_mensaje( &mut self, mensaje: MensajeCliente, ) -> Result<Respuesta, ErrorEnThread>

Se encarga de ejecutar lo que el cliente pidio y devolver una Respuesta adecuada.

Para esto pide el lock de la calculadora para asi poder ejecutar la peticion del cliente sobre esta.

En caso de algun error durante estas operaciones se retorna un ErrorEnThread

fn enviar_respuesta( &mut self, respuesta: Respuesta, ) -> Result<(), ErrorEnThread>

Envia al cliente una Respuesta por medio del writer_stream.

Si ocurre algun tipo de error se retorna un ErrorEnThread