Module server::controller

This module is the controller of the system. It receives start and stop commands from the HTTP module (when the HTTP module receives start and stop requests). When the server is started, it will read segments from the offline file and send them to each of the connected TCP client managers.

The system as a whole is built around using futures in a green threaded event driven environment using Tokio.

When the controller is created it checks whether it should create a cache or not. If it should create the cache it loops through the CSV file and writes the values out in network order into 60 files, one file for each channel. In other words, the cache files are files that contain the data to be sent directly on the network for each channel.

When the controller is run it runs a loop that does the following. First it checks for config updates (from the HTTP server). If the server was not running it will sleep until started. Then it proceeds to collect a segment according to the configuration parameters for the server. Lastly it tries to send the segment to all TCP client managers with their TX. It records which TXes fail at sending. This may be because the client has disconnected so the RX side does not exist anymore, or it may be because the channel is full because of backpressure. In either case these clients are dropped. If there is a corresponding RX side, this will fail closing the TCP connection cleanly.

Re-exports

use byteorder::BigEndian;

use byteorder::ByteOrder;

use bytes::BufMut;

use bytes::Bytes;

use bytes::BytesMut;

use csv::ReaderBuilder;

use futures::sync::oneshot;

use std;

use std::fs;

use std::thread;

use std::io::BufReader;

use std::io::Read;

use std::io::Seek;

use std::io::SeekFrom;

use std::io::Write;

use std::time::Duration;

use std::time::Instant;

use tcp::Clients;

Structs

Config
Represents the server state and is configured using the POST /start.

Controller
The controller data struct.

Sample
Represents a single line in the CSV file.

Enums

Command
The command to send on the Command channel.

Constants

_IMPL_DESERIALIZE_FOR_Config

_IMPL_DESERIALIZE_FOR_Sample

Functions

duration_from_micros
Creates a Duration object from microseconds.

Type Definitions

CommandRx
The type for the receiving side of the Command channel.

CommandTx
The type for the sending side of the Command channel.

Running
Represents a running state.