Struct server::controller::Controller

pub struct Controller {
    command_rx: Receiver<(Command, Sender<()>)>,
    clients: Arc<Mutex<HashMap<SocketAddr, Sender<Bytes>>>>,
    config: Option<Config>,
    samples: Vec<BufReader<File>>,
    last_segment_finished: Instant,
}

The controller data struct.

Fields
command_rx: Receiver<(Command, Sender<()>)>

Receives commands from the HTTP module.

clients: Arc<Mutex<HashMap<SocketAddr, Sender<Bytes>>>>

List of TCP clients to send CSV data to.

config: Option<Config>

The global server config.

samples: Vec<BufReader<File>>

A vector containing a buffered file reader (performance) for each channel.

last_segment_finished: Instant

Keep track of when the last segment was sent so we know when to send the next one.

Methods

impl Controller

The controller creates a cache, receives commands from the HTTP module, and sends data from the cache to the TCP clients.

pub fn new(
command_rx: Receiver<(Command, Sender<()>)>,
clients: Arc<Mutex<HashMap<SocketAddr, Sender<Bytes>>>>,
filename: Option<String>
) -> Controller

fn update_config(&mut self)

Update the global config into the local config variable.

This will block if the server is stopped, else it will not block.

fn sleep_until(&self, instant: Instant)

Simple implementation of a sleep until function.

fn collect_segment(&mut self) -> Bytes

Collect a segment from the cache.

If the cache is completely consumed, start over at the top of the cache files.

pub fn run(self)

Run the controller loop.

This will kick any clients who are not receiving fast enough.