server::domain

Enum GameCommand

pub enum GameCommand {
    HostGame {
        user_id: UserId,
        game_id: GameId,
    },
    JoinGame {
        user_id: UserId,
    },
    PlayComputer {
        user_id: UserId,
        game_id: GameId,
    },
    CutForDeal {
        player: Player,
    },
    StartGame {
        player: Player,
    },
    DiscardCards {
        player: Player,
        cards: Vec<Card>,
    },
    PlayCard {
        player: Player,
        card: Card,
    },
    Go {
        player: Player,
    },
    ScorePone {
        player: Player,
    },
    ScoreDealer {
        player: Player,
    },
    ScoreCrib {
        player: Player,
    },
    StartNextRound {
        player: Player,
    },
}

Commands that drive the game state machine.

Each variant represents an action a player can take.

Variants

HostGame

User wants to create a new game and become its host. The game will require a guest to join in order to continue.

Fields

user_id: UserId

User hosting the game.

game_id: GameId

The new game identity.

JoinGame

User wants to join the game has a guest.

Fields

user_id: UserId

User hosting the game.

PlayComputer

Client wants to start a single-player game against the computer.

Fields

user_id: UserId

User creating the game.

game_id: GameId

The new game identity.

CutForDeal

Player cuts the deck to determine who deals first.

Fields

player: Player

The player identity (essentially host or guest)

StartGame

Player signals they are ready for the game to start. This is a Pending acknowledgment on the CutForDeal cuts.

Fields

player: Player

The player identity (essentially host or guest)

DiscardCards

Player discards two cards to the crib.

Fields

player: Player

The player identity (essentially host or guest)

cards: Vec<Card>

The cards being discarded.

PlayCard

Player plays a card during the Playing (or Pegging) phase.

Fields

player: Player

The player identity (essentially host or guest)

card: Card

The card to be played.

Go

Player declares go during the Playing (or Pegging) phase.

Fields

player: Player

The player identity (essentially host or guest)

ScorePone

Player requests scoring of the non-dealer’s hand. This is a Pending acknowledgment at the end of the Playing phase.

Fields

player: Player

The player identity (essentially host or guest)

ScoreDealer

Player requests scoring of the dealer’s hand. This is a Pending acknowledgment at the end of the ScoringPone phase.

Fields

player: Player

The player identity (essentially host or guest)

ScoreCrib

Player requests scoring of the dealer’s crib. This is a Pending acknowledgment at the end of the ScoringDealer phase.

Fields

player: Player

The player identity (essentially host or guest)

StartNextRound

Player requests that the next Discarding round should start. This is a Pending acknowledgment at the end of the ScoringCrib phase.

Fields

player: Player

The player identity (essentially host or guest)