server::domain

Enum GameEvent

pub enum GameEvent {

    LobbyGameCreated {
        game_id: GameId,
        host: UserId,
        name: String,
    },
    LobbyGameJoined {
        guest: UserId,
    },
    ComputerGameCreated {
        game_id: GameId,
        host: UserId,
        guest: UserId,
        name: String,
    },
    CutForDealMade {
        player: Player,
        cut: Card,
    },
    GameStarted {
        player: Player,
    },
    CutForDealTied,
    CutForDealDecided {
        dealer: Dealer,
    },
    HandDealt {
        player: Player,
        hand: Hand,
    },
    CardsDiscarded {
        player: Player,
        cards: Vec<Card>,
    },
    StarterSelected {
        cut: StarterCut,
        pegging: Pegging,
    },
    CardPlayed {
        player: Player,
        card: Card,
        pegging: Pegging,
    },
    GoCalled {
        player: Player,
        pegging: Pegging,
    },
    PoneScored {
        player: Player,
        pegging: Pegging,
    },
    DealerScored {
        player: Player,
        pegging: Pegging,
    },
    CribScored {
        player: Player,
        pegging: Pegging,
    },
    NextRoundStarted {
        player: Player,
    },

}

Domain events which represent the single source of truth for game history and are persisted in the event store. The current game state is derived by folding them in order.

Variants

LobbyGameCreated

A user created a lobby game and is waiting for another user to join

Fields

game_id: GameId

The identity of the game that was created.

host: UserId

The user that created the game.

name: String

The name assigned to the game. It’s possible this may be the same as another game; the game_id provides the unique id.

LobbyGameJoined

A user joined a game that was previously a lobby game.

Fields

guest: UserId

The user that joined the game.

ComputerGameCreated

A user a computer game.

Fields

game_id: GameId

The identity of the game that was created.

host: UserId

The user that created the game.

guest: UserId

The computer “user” that will play against the host.

name: String

The name assigned to the game. It’s possible this may be the same as another game; the game_id provides the unique id.

CutForDealMade

A player cut a card for the choice of initial dealer.

Fields

player: Player

The player identity (essentially host or guest)

cut: Card

The card that was cut.

GameStarted

A player has acknowledged the cut for deal. When both players have acknowledge the cards will be dealt and discarding begin.

Fields

player: Player

The player identity (essentially host or guest)

CutForDealTied

Both plays have cut for deal (and acknowledged the cut); the cuts were the same rank, so the cut needs to be redrawn.

CutForDealDecided

Both plays have cut for deal (and acknowledged the cut); the cuts were different, so the dealer is selected and the deal can progress.

Fields

dealer: Dealer

The selected dealer.

HandDealt

A hand has been dealt to the player, following CutForDealDecided or NextRoundStarted.

Fields

player: Player

The player identity (essentially host or guest)

hand: Hand

The hand that has been dealt.

CardsDiscarded

A player has discarded cards to the crib.

Fields

player: Player

The player identity (essentially host or guest)

cards: Vec<Card>

The cards that were discarded.

StarterSelected

Both players have discarded cards to the crib and the “Starter” card selected.

Fields

cut: StarterCut

The starter card.

pegging: Pegging

Record HisHeels.

CardPlayed

A player has played a card during the Playing (or Pegging) phase.

Fields

player: Player

The player identity (essentially host or guest)

card: Card

The card that was played.

pegging: Pegging

Points awarded during pegging (may include 31 for 2).

GoCalled

A player has called Go during the Playing (or Pegging) phase.

Fields

player: Player

The player identity (essentially host or guest)

pegging: Pegging

Points awarded for the Go.

PoneScored

A player has acknowledged End of Plays; when both players have acknowledged then the Pone’s hand will be scored.

Fields

player: Player

The player identity (essentially host or guest)

pegging: Pegging

The pone’s hand scores.

DealerScored

A player has acknowledged the Pone Score; when both players have acknowledged then the Dealer’s hand will be scored.

Fields

player: Player

The player identity (essentially host or guest)

pegging: Pegging

The dealer’s hand scores.

CribScored

A player has acknowledged the Dealer Score; when both players have acknowledged then the Dealer’s crib will be scored.

Fields

player: Player

The player identity (essentially host or guest)

pegging: Pegging

The dealer’s crib scores.

NextRoundStarted

A player has acknowledged the Crib Score; when both players have acknowledged then the next round will start, swapping dealers.

Fields

player: Player

The player identity (essentially host or guest)