server::domain

Type Alias Deck

pub type Deck = Pile<DeckType>;

The deck contains all cards at the start of the game and is drawn from during dealing and for the starter cut.

Aliased Type

pub struct Deck { /* private fields */ }

Implementations

impl Deck

pub fn shuffled_pack() -> Self

pub fn cut(&mut self) -> Card

pub fn deal(&mut self, count: usize) -> Vec<Hand>