Function server::handle_connection

pub(crate) fn handle_connection(
    stream: TcpStream, 
    data: &mut HashMap<String, String>
) -> Result<()>

Handle user connection

Takes user’s stream as first argument and hashmap in which all keys are saved as second Transforms json format to String -> checks request -> sends answer based on rules: Load request: { Answer: { “response_status”: “success”, “requested_key”: “some_key”, “requested_hash”: “0b672dd94fd3da6a8d404b66ee3f0c83”, } Store request: { “request_type”: “store”, “key”: “some_key”, “hash”: “0b672dd94fd3da6a8d404b66ee3f0c83” } Answer: { “response_status”: “success”, }