server::messages::serialization
Function deserialize_message

pub fn deserialize_message(response: &str) -> Result<Message>

Deserializes a JSON string into a Message enum variant.

Arguments

response - A string slice that holds the JSON representation of the message.

Returns

A Result<Message, serde_json::Error> which is Ok containing the Message enum variant if deserialization is successful, or an Err containing the serde_json::Error if it fails.

Note

Currently, it only supports deserializing FragmentRequest. Other types like FragmentTask and FragmentResult are marked with todo!().