Module server::models::changelog

Changelog-specific models.

Structs

BannedTimeDetails

Details on a user’s banned/unverified runs for the admin display page

CalcValues

Values that we return after checking if a score is valid to be added to the database.

Changelog

One-to-one struct for changelog data.

ChangelogInsert

All changelog data except for the ID, for table insertion.

ChangelogPage

Indlues additional information from joins that includes details like map name, username and profile image.

ChangelogQueryParams

All the accepted query parameters for the SubmissionChangelog page.

ChangelogSearchQuery

Query parameters for searching for a given

EvidenceRequirements

One-to-one struct for evidence_requirements

HistoryParams

Wrapper to allow queries to include map_id, profile_number and optional cat_id.

MapScoreDate

Map ID & Name, score and timestamp for a given score.

NumScores

Used to count the number of scores per-user.

NumUpdatePerMap

Representation of the number of World Records per Map for a given map.

Recap

Struct for the “Recap”, taken from NeKz’s recap bot on the Discord server.

ScoreDeltaComparison

Used to represent users and their score deltas on a given map.

ScoreLookup

Used to lookup information on a specific score.

ScoreParams

Wrapper to send a profile number as a search result

SubmissionChangelog

Fields for a submission to the changelog