Struct server::models::models::Users

pub struct Users {
    pub profile_number: String,
    pub board_name: Option<String>,
    pub steam_name: Option<String>,
    pub banned: bool,
    pub registered: i32,
    pub avatar: Option<String>,
    pub twitch: Option<String>,
    pub youtube: Option<String>,
    pub title: Option<String>,
    pub admin: i32,
    pub donation_amount: Option<String>,
    pub discord_id: Option<String>,
}

One-to-one struct for user data.

Fields
profile_number: Stringboard_name: Option<String>steam_name: Option<String>banned: boolregistered: i32avatar: Option<String>twitch: Option<String>youtube: Option<String>title: Option<String>admin: i32donation_amount: Option<String>discord_id: Option<String>
Implementations

impl Users

pub async fn get_user(
pool: &PgPool,
profile_number: String
) -> Result<Option<Users>>

Returns user information

pub async fn get_user_data(
pool: &PgPool,
profile_number: &str
) -> Result<Option<UsersPage>>

Gets a user’s avatar and user_name/board_name (favors board_name)

pub async fn check_board_name(
pool: &PgPool,
nick_name: String
) -> Result<Option<Vec<String>>>

Pattern match on a given string to find similar names (supports board/steam names).

pub async fn get_banned(pool: &PgPool) -> Result<Vec<String>>

Returns a list of all banned player’s profile_numbers.

pub async fn get_banned_display(
pool: &PgPool
) -> Result<Option<Vec<UsersDisplay>>>

Returns a list of all banned player’s as a UsersDisplay object.

pub async fn check_banned(pool: &PgPool, profile_number: String) -> Result<bool>

Returns the boolean flag associated with the user in the boards, if Err, assumed User does not exist.

pub async fn get_title(
pool: &PgPool,
profile_number: String
) -> Result<Option<String>>

Returns the title associated with the user (CAN BE NONE)

pub async fn get_socials(
pool: &PgPool,
profile_number: String
) -> Result<Option<Socials>>

Returns the social media informatio associated with a given user’s profile_number

pub async fn get_admin_for_user(
pool: &PgPool,
profile_number: String
) -> Result<Option<i32>>

Returns the admin information associated with the user.

pub async fn get_all_admins(
pool: &PgPool,
admin_value: i32
) -> Result<Option<Vec<UsersDisplay>>>

Returns UsersDisplay for all admins Usage: admin_value = 0 -> Non-admin user admin_value = 1 -> Standard admin admin_value = 2 -> Shadow admin (Has admin permissions, is not publically listed) (Typically reserved for former admins, trusted players) admin_value = 3 -> Developer admin (Has admin permissions as an activen developer only)

pub async fn get_donators(pool: &PgPool) -> Result<Option<Vec<Users>>>

Returns all users that have donated to the board. Ordered by highest amount.

pub async fn get_profile(
pool: &PgPool,
profile_number: &String
) -> Result<Option<ProfileData>>

pub async fn insert_new_users(pool: &PgPool, new_user: Users) -> Result<bool>

Inserts a new user into the databse

pub async fn update_existing_user(
pool: &PgPool,
updated_user: Users
) -> Result<bool>

pub async fn delete_user(pool: &PgPool, profile_number: String) -> Result<bool>