Module server::net::types

Structs

ClientErrMsg
Struct to send the kind of error and error message to the client

DataSet
Representation of a ResultSet with its useful functions to get data.

Greeting
This is the first packet being sent by the server after the TCP connection is established.

Login
The client responds with this packet to a Greeting packet, finishing the authentication handshake.

Enums

Command
Sent by the client to the server.

PkgType
Code numeric value sent as first byte

Functions

preprocess
Sort the Vec data into DataSet for further use.