Struct server::net::types::DataSet

pub struct DataSet {
    // some fields omitted
}

Representation of a ResultSet with its useful functions to get data.

Methods

impl DataSet

fn get_col_cnt(&self) -> usize

fn data_empty(&self) -> bool

fn metadata_empty(&self) -> bool

fn get_col_idx(&self, name: String) -> Option<usize>

fn get_col_name(&mut self, idx: usize) -> Option<&str>

fn get_type_by_name(&mut self, name: String) -> Option<SqlType>

fn get_is_primary_by_name(&mut self, name: String) -> Option<bool>

fn get_allow_null_by_name(&mut self, name: String) -> Option<bool>

fn get_description_by_name(&mut self, name: String) -> Option<&str>

fn get_type_by_idx(&mut self, idx: usize) -> Option<SqlType>

fn get_is_primary_key_by_idx(&mut self, idx: usize) -> Option<bool>

fn get_allow_null_by_idx(&mut self, idx: usize) -> Option<bool>

fn get_description_by_idx(&mut self, idx: usize) -> Option<&str>

fn next_int_by_idx(&mut self, idx: usize) -> Option<i32>

Return next data entry. next() has to be called first it initialize the pointer

fn next_bool_by_idx(&mut self, idx: usize) -> Option<bool>

Return next data entry. next() has to be called first it initialize the pointer

fn next_char_by_idx(&mut self, idx: usize) -> Option<String>

Return next data entry. next() has to be called first it initialize the pointer

fn next_int_by_name(&mut self, name: String) -> Option<i32>

Return next data entry. next() has to be called first it initialize the pointer

fn next_bool_by_name(&mut self, name: String) -> Option<bool>

Return next data entry. next() has to be called first it initialize the pointer

fn next_char_by_name(&mut self, name: String) -> Option<String>

Return next data entry. next() has to be called first it initialize the pointer

fn first(&mut self)

Set the data pointer before the first entry (pos = -1). next() has to be called first to start a new next... - loop

fn last(&mut self)

Set the data pointer after the last entry . previous() has to be called first to start a new backward loop

fn next(&mut self) -> bool

Move the pointer to the next line. Return false if end of data, else true.

fn previous(&mut self) -> bool

Move the pointer to the previous line. Return false if end of data, else true.