Module server::packet

Packet

Different packet management and processing.

Enums

Packet

Type of recognized packets.

Constants

CONNECTION_IDENTIFIER_REFUSED

CONNECTION_PROTOCOL_REJECTED

CONNECTION_USER_OR_PASS_REFUSED

INCORRECT_SERVER_CONNECTION

MQTT_NAME

MQTT_VERSION

SUCCESSFUL_CONNECTION

Functions

bytes2string

Convert bytes to UTF-8 string.

change_subscription

Inform coordinator that the client has changed a subscription.

close_streams

Close the stream with the client.

inform_client_disconnect_to_coordinator

Inform coordinator that the client sent a Disconnect packet.

inform_new_connection

Inform the coordinator that a new client has connected.

make_connection

Process Connection packet.

make_publication

Send new Publish packet to the coordinator.

read_packet

Read entire packet from stream and execute the function according to type.

remove_from_client_publishes

send_connection_error

Send a Connack packet to the client with the connection error code.

send_pingresp

Send Pingresp to the client.

send_publication_results

Send Puback packet to the client.

user_and_password_correct

Check if username and password are valid.

verify_protocol_name

Check if protocol name is the required for a new connection.

verify_version_protocol

Check if protocol version is the required for a new connection.