Module server::parse

This module contains functions and types for parsing SQL queries.

Reexports

pub use self::parser::Parser;

Modules

ast

lex

parser

token

Structs

Span
Represents a substring in the query string in byte indices.

Functions

parse
Main function of this module: Takes a sql query as string and returns the parsed AST.