server

Module request_handler

Structs

InternalResponse

Response

Functions

change_asset_color

confirm_asset_ability

divest_asset

draw_card

end_turn

fire_character

get_bonus_cash

minus_into_plus

pay_banker

play_card

put_back_card

redeem_liability

resync

Facilitates a client resync by providing a packet containing the full gamestate Contains data specific to the current gamestate

select_character

select_divest_asset

select_issue_liability

silver_into_gold

start_game

swap_with_deck

swap_with_player

terminate_credit_character

unselect_divest_asset

unselect_issue_liability

use_ability