Struct server::storage::bstar::Bstar

pub struct Bstar<T: Debug + PartialOrd + KnownSize> {
    pub root: u64,
    pub elementcount: u64,
    pub order: u64,
    pub freeaddr: u64,
    pub eof: u64,
    pub allowduplicates: bool,
    pub target: String,
    // some fields omitted
}

Fields

root

elementcount

order

freeaddr

eof

allowduplicates

target

Methods

impl<T: KnownSize + PartialOrd + Clone + Debug> Bstar<T>

fn delete(name: &str) -> Result<()>

fn load(name: &str) -> Result<Bstar<T>>

Loads a Bstar object from the specified name/path

fn create(name: &str, target: &str, order: u64) -> Result<Bstar<T>>

Creates a new permanent Bstar object. target is the name of the table holding the data records, order*2 specifies the maximal amount of keys stored per node in the tree

fn reset(&mut self) -> Result<()>

resets the tree to 0 elements

fn get_root(&mut self) -> Result<Bnode<T>>

returns the rootnode of the tree

fn debug_print(&mut self) -> Result<()>

prints a debug version of the tree

fn lookup_keyaddr(&mut self, key: T) -> Result<Option<KeyAddr<T>>>

searches for key in the tree and returns the KeyAddr object or None

fn insert_keyaddr(&mut self, key: KeyAddr<T>) -> Result<bool>

inserts a keyaddr object

fn delete_keyaddr(&mut self, key: T) -> Result<Option<KeyAddr<T>>>

searches for the key and deletes and returns the keyaddr object

fn iter(&mut self) -> Bterator<T>

returns an iterator for the elements of the tree

fn iter_options(&mut self, direction: IterDirection, key: Option<IterOption<T>>) -> Bterator<T>

returns a specific iterator moving in the specified direction if a key is given to the function with the Option to include or exclude it, the iterator will do so. It does not matter if the key is in the tree or not.