Enum server::storage::DatabaseError

pub enum DatabaseError {
    Io(Error),
    BinEn(EncodingError),
    BinDe(DecodingError),
    Byteorder(Error),
    WrongMagicNmbr,
    Engine,
    LoadDataBase,
    RemoveColumn,
    AddColumn,
}

A database table

Through this type, you can retreive certain meta information about the table (like column names, column types, storage engine, ...). It's access method locks the table globally and returns a storage engine to access the table data.

Variants

Io

BinEn

BinDe

Byteorder

WrongMagicNmbr

Engine

LoadDataBase

RemoveColumn

AddColumn