Struct server::storage::Rows

pub struct Rows<B: Write + Read + Seek> {
    pub columns: Vec<Column>,
    pub column_offsets: Vec<u64>,
    // some fields omitted
}

Fields

columns

column_offsets

Methods

impl<B: Write + Read + Seek> Rows<B>

Represents the lines read from file.

fn new(data_src: B, columns: &[Column]) -> Rows<B>

fn next_row<W: Write>(&mut self, target_buf: &mut W) -> Result<u64, Error>

reads the next row, which is not marked as deleted and writes the data into target_buf returns the bytes read or an Error otherwise. Returns Error:EndOfFile if no next row could be read.

fn reset_pos(&mut self) -> Result<u64, Error>

sets position before the first line

fn add_row(&mut self, data: &[u8]) -> Result<u64, Error>

writes a new row into buf, returns bytes written

fn reorganize(&mut self) -> Result<u64, Error>

Reorganizes the current instance Removes all rows which are marked as deleted. Returns the new size of the reorganized object.

fn get_value(&self, row_data: &[u8], column_index: usize) -> Result<Vec<u8>, Error>

returns the value of the column_index' column of the current row returns Error::InvalidState if no current row exists

fn get_column(&self, index: usize) -> &Column

returns the columns

fn insert_row_without_primary(&mut self, row_data: &[u8]) -> Result<u64, Error>

Inserts a new row with row_data. Does not check if the primary key exists. Returns the number of rows_inserted.

fn insert_row(&mut self, row_data: &[u8]) -> Result<u64, Error>

Inserts a new row with row_data. Returns the number of rows inserted.

fn delete(&mut self, column_index: usize, value: (&[u8], Option<usize>), comp: CompType) -> Result<u64, Error>

deletes rows which fulfills a constraint return rows deleted

fn modify(&mut self, constraint_column_index: usize, constraint_value: (&[u8], Option<usize>), comp: CompType, values: &[(usize, &[u8])]) -> Result<u64, Error>

Updates all rows fulfilling the constraint. constraint_column_index: index of the rows whose value is compared to constraint_value. values[(column_index: usize, new_value: &[u8])]: Slice of tuples The first value of the tuple contains the index of the column to be updated. The second value contains the new value for the column. returns the numer of rows updated. Panics if a row could not be updated. If modify fails, the updated row will be lost.

fn lookup(&mut self, column_index: usize, value: (&[u8], Option<usize>), comp: CompType) -> Result<Rows<Cursor<Vec<u8>>>, Error>

Returns an new Rows object with all rows which fulfill the constraint. column_index: index of the column whose value should match value comp: defines how the value of the column and value should be compared.

fn full_scan(&mut self) -> Result<Rows<Cursor<Vec<u8>>>, Error>

scans the entire file returns all rows which are not deleted

fn is_empty(&mut self) -> Result<bool, Error>

checks if object is containing rows returns bool on success else Error

fn get_next_row(&mut self, column_index: usize, value: (&[u8], Option<usize>), comp: CompType) -> Result<Vec<u8>, Error>

moves the cursor from the current position the row after the first row which fulfills the constraint. Returns the first row which fulfilly the condition.

fn to_result_set(&mut self) -> Result<ResultSet, Error>

Returns a new ResultSet containing all rows of the current object