Struct server::storage::Table

pub struct Table<'a> {
    pub name: String,
    // some fields omitted
}

Table struct that contains the table information

Fields

name

Methods

impl<'a> Table<'a>

fn new<'b>(database: &'b Database, name: &str, columns: Vec<Column>, engine_id: u8) -> Table<'b>

Creates new table object Returns Table

fn save(&self) -> Result<(), DatabaseError>

Saves the table with a identification number in table file Returns DatabaseError on fail else Nothing

fn delete(&self) -> Result<(), DatabaseError>

Deletes the .tbl files Returns DatabaseError on fail if path points to a directory, if the user lacks permissions to remove the file, or if some other filesystem-level error occurs.

fn columns(&self) -> &[Column]

Returns columns of table as array

fn add_column(&mut self, name: &str, sql_type: SqlType) -> Result<(), DatabaseError>

Adds a column to the tabel Returns name of Column or on fail DatabaseError

fn remove_column(&mut self, name: &str) -> Result<(), DatabaseError>

Removes a column from the table Returns name of Column or on fail DatabaseError

fn create_engine(self) -> Box<Engine + 'a>

Creates an engine for Table Returns Box