server
Struct Server

pub struct Server { /* private fields */ }

Represents a server node that implements both the NetworkNode and HostNode traits. It is responsible for handling network packets, simulation events, file requests, chat, and more.

Implementations

impl Server

pub fn new( id: NodeId, controller_send: Sender<DroneEvent>, controller_recv: Receiver<ServerCommand>, packet_recv: Receiver<Packet>, packet_send: HashMap<NodeId, Sender<Packet>>, ) -> Self

Creates a new server instance with the specified parameters.