Struct server::tcp::Server

pub struct Server {
    clients: Arc<Mutex<HashMap<SocketAddr, Sender<Bytes>>>>,
    listener: TcpListener,
}

Represents the state of the server.

Fields
clients: Arc<Mutex<HashMap<SocketAddr, Sender<Bytes>>>> listener: TcpListener
Methods

impl Server

pub fn bind(addr: &SocketAddr) -> Server

Binds the TCP server to the specified address.

pub fn get_clients(&self) -> Arc<Mutex<HashMap<SocketAddr, Sender<Bytes>>>>

Gets a reference counted reference to the clients HashMap.

pub fn run(self)

Runs the server, setting up the green threaded event loop.

The TCP client manager (event loop) will read any data on the Client channel and output it onto the TCP connection.