Module server::tools::cache

Caching for endpoints Caching implementations specific to the board.

Currently built off std::collections::HashMap.

Accessing in endpoints.

use crate::tools::cache::CacheState;
use actix_web::{get, web, HttpResponse, Responder};

#[get("/test")]
async fn test(cache: web::Data<CacheState>) -> impl Responder {
    let cache = cache.into_inner(); // Extracts the CacheState from the [actix_web::web::Data] wrapper
    // Access the default category ids.
    let map_id = "47458";
    let def_cat_id = cache.default_cat_ids[map_id];

    // Check the current cache state for points.
    let state_data = &mut cache.current_state.lock().await; // We use &mut here so that we can change the value accordingly.
    // Check the cache state for "coop_previews"
    // Check `cached_endpoints` in CacheState::new() for which endpoints are current cached.
    let is_cached = state_data.get_mut("coop_previews").unwrap();
    // Derefernce the bool and match accordingly.
    if !*is_cached {
        // Assume we cache here, to see full example check [crate::server::api::v1::handlers::coop::get_cooperative_preview]
        *is_cached = true;
    }

    // Check points which maps as follows:
    // Points descriptor str -> profile_number -> [crate::models::points::Points]
    let points_id = "points_sp";
    let points_hm = &mut cache.points.lock().await;
    // This gives us a &mut HashMap that maps profile_number to Points struct
    // For full example see [server::api::v1::handlers::points::post_points_sp]
    let points_cache = points_hm.get_mut(points_id).unwrap();

    // Use Ranks
    let profile_number = "76561198135023038".to_string();
    let ranks = &mut cache.ranks.lock().await;
    // Accessing current_ranks field, then using as a hashmap with the mapping from map_id -> rank
    let user = ranks
        .current_ranks
        .entry(profile_number)
        .or_insert_with(HashMap::new);

    HttpResponse::Ok().body("test")
}

Structs

CacheState

Holds a thread-sharable hashmap that we use to control cache invalidation.

Ranks

Cache for the current ranks all players have within the top X scores (defined by crate::tools::config::ProofConfig)

Constants

COOP_PREVIEWS

POINTS_1

POINTS_2

POINTS_3

POINTS_4

POINTS_5

POINTS_6

POINTS_7

POINTS_8

POINTS_9

POINTS_10

POINTS_11

POINTS_12

POINTS_13

POINTS_14

POINTS_15

POINTS_COOP

POINTS_OVERALL

POINTS_SP

SP_PREVIEWS

Functions

read_from_file

Reads data from a file for any type that implements serde::Deserialize

write_to_file

Writes data to a file if the type implements serde::Serialize