Crate server

Modules

codec

server
ChatServer is an actor. It maintains list of connection client session. And manages available rooms. Peers send messages to other peers in same room through ChatServer.

session
ClientSession is an actor, it manages peer TCP connection and proxies commands from peer to ChatServer.

Structs

Server
Define TCP server that will accept incoming TCP connection and create chat actors.

TcpConnect

Functions

main