Crate server

Implementation of a server using Pencil and the FenixEDU API with Heroku support. There are seven routes defined for the API (/api/).

For room management there are only three relevant fields from the FenixEDU API response: name, containedSpaces and capacity. A space is considered a room when containedSpaces has zero elements.

REST API

GET

spaces => Returns the top level spaces from the FenixEDU API;

id/<id> => Returns the list of contained spaces, name and capacity when relevant inside each id;

rooms => Returns the rooms available to check-in and check-out of in the DB;

path/<my_path> => Returns the contained spaces, name and capacity when applicable for the specified hierarchical path.

POST

create_user => Creates a user in the database;

create_room => Adds a room to the database. A room exists when the contained_space list is empty;

check_in => Adds a user to a specified room.

DELETE

check_out => Removes a user from a specified room.