Crate server

Modules

auth
Module for authenticating users and check privileges

conn
Contains the entry point code for handling an incoming connection.

logger
Logging functionality

net
The network api

parse
This module contains functions and types for parsing SQL queries.

query
Query excecution moduleT

storage
Storage Engine trait and several implementations

Structs

Config
A struct for managing configurations

Functions

listen
Listens for incoming TCP streams