Crate server

Modules

db_queries ğŸ”’

Enums

ServerError

Functions

chat ğŸ”’

chat implement main processing loop.

listen_and_accept ğŸ”’

listen_and_accept take care of connection of new client connections.

send_to_everyone_else ğŸ”’

send_to_everyone_else process sending of message to every client other to the message sender.

start_server

start_server is entrypoint of server. It starts main processing loop in a separate thread while main thread keep8s track on managing new client connections.

Type Aliases

ClientMap ğŸ”’

Clients ğŸ”’